refactor: Request,Response 디렉터리 분리
