+++ src/main/java/hanium/modic/backend/web/user/controller/UserController.java
@@ -15,12 +15,12 @@ import hanium.modic.backend.common.response.AppResponse;
 import hanium.modic.backend.domain.user.entity.UserEntity;
 import hanium.modic.backend.domain.user.service.UserCoinService;
 import hanium.modic.backend.domain.user.service.UserService;
-import hanium.modic.backend.web.user.dto.TransferCoinsRequest;
-import hanium.modic.backend.web.user.dto.UpdateUserNameRequest;
-import hanium.modic.backend.web.user.dto.UpdateUserPasswordRequest;
-import hanium.modic.backend.web.user.dto.UserCreateRequest;
-import hanium.modic.backend.web.user.dto.UserCreateResponse;
-import hanium.modic.backend.web.user.dto.UserInfoResponse;
+import hanium.modic.backend.web.user.dto.request.TransferCoinsRequest;
+import hanium.modic.backend.web.user.dto.request.UpdateUserNameRequest;
+import hanium.modic.backend.web.user.dto.request.UpdateUserPasswordRequest;
+import hanium.modic.backend.web.user.dto.request.UserCreateRequest;
+import hanium.modic.backend.web.user.dto.response.UserCreateResponse;
+import hanium.modic.backend.web.user.dto.response.UserInfoResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import jakarta.validation.Valid;
