+++ src/main/java/hanium/modic/backend/web/user/dto/request/UserCreateRequest.java
@@ -1,4 +1,4 @@
-package hanium.modic.backend.web.user.dto;
+package hanium.modic.backend.web.user.dto.request;
 
 import org.hibernate.validator.constraints.Length;
